自动初始化单独建仓 为 命令行做基础
