+++ server/model/sys_casbin.go
@@ -6,145 +6,3 @@ type CasbinModel struct {
 	Path        string `json:"path" gorm:"column:v1"`
 	Method      string `json:"method" gorm:"column:v2"`
 }
-
-func CasbinModelData() []CasbinModel {
-	return []CasbinModel{
-		{"p", "888", "/base/login", "POST"},
-		{"p", "888", "/base/register", "POST"},
-		{"p", "888", "/api/createApi", "POST"},
-		{"p", "888", "/api/getApiList", "POST"},
-		{"p", "888", "/api/getApiById", "POST"},
-		{"p", "888", "/api/deleteApi", "POST"},
-		{"p", "888", "/api/updateApi", "POST"},
-		{"p", "888", "/api/getAllApis", "POST"},
-		{"p", "888", "/authority/createAuthority", "POST"},
-		{"p", "888", "/authority/deleteAuthority", "POST"},
-		{"p", "888", "/authority/getAuthorityList", "POST"},
-		{"p", "888", "/authority/setDataAuthority", "POST"},
-		{"p", "888", "/authority/updateAuthority", "PUT"},
-		{"p", "888", "/authority/copyAuthority", "POST"},
-		{"p", "888", "/menu/getMenu", "POST"},
-		{"p", "888", "/menu/getMenuList", "POST"},
-		{"p", "888", "/menu/addBaseMenu", "POST"},
-		{"p", "888", "/menu/getBaseMenuTree", "POST"},
-		{"p", "888", "/menu/addMenuAuthority", "POST"},
-		{"p", "888", "/menu/getMenuAuthority", "POST"},
-		{"p", "888", "/menu/deleteBaseMenu", "POST"},
-		{"p", "888", "/menu/updateBaseMenu", "POST"},
-		{"p", "888", "/menu/getBaseMenuById", "POST"},
-		{"p", "888", "/user/changePassword", "POST"},
-		{"p", "888", "/user/getUserList", "POST"},
-		{"p", "888", "/user/setUserAuthority", "POST"},
-		{"p", "888", "/user/deleteUser", "DELETE"},
-		{"p", "888", "/fileUploadAndDownload/upload", "POST"},
-		{"p", "888", "/fileUploadAndDownload/getFileList", "POST"},
-		{"p", "888", "/fileUploadAndDownload/deleteFile", "POST"},
-		{"p", "888", "/casbin/updateCasbin", "POST"},
-		{"p", "888", "/casbin/getPolicyPathByAuthorityId", "POST"},
-		{"p", "888", "/casbin/casbinTest/:pathParam", "GET"},
-		{"p", "888", "/jwt/jsonInBlacklist", "POST"},
-		{"p", "888", "/system/getSystemConfig", "POST"},
-		{"p", "888", "/system/setSystemConfig", "POST"},
-		{"p", "888", "/customer/customer", "POST"},
-		{"p", "888", "/customer/customer", "PUT"},
-		{"p", "888", "/customer/customer", "DELETE"},
-		{"p", "888", "/customer/customer", "GET"},
-		{"p", "888", "/customer/customerList", "GET"},
-		{"p", "888", "/autoCode/createTemp", "POST"},
-		{"p", "888", "/autoCode/getTables", "GET"},
-		{"p", "888", "/autoCode/getDB", "GET"},
-		{"p", "888", "/autoCode/getColume", "GET"},
-		{"p", "888", "/sysDictionaryDetail/createSysDictionaryDetail", "POST"},
-		{"p", "888", "/sysDictionaryDetail/deleteSysDictionaryDetail", "DELETE"},
-		{"p", "888", "/sysDictionaryDetail/updateSysDictionaryDetail", "PUT"},
-		{"p", "888", "/sysDictionaryDetail/findSysDictionaryDetail", "GET"},
-		{"p", "888", "/sysDictionaryDetail/getSysDictionaryDetailList", "GET"},
-		{"p", "888", "/sysDictionary/createSysDictionary", "POST"},
-		{"p", "888", "/sysDictionary/deleteSysDictionary", "DELETE"},
-		{"p", "888", "/sysDictionary/updateSysDictionary", "PUT"},
-		{"p", "888", "/sysDictionary/findSysDictionary", "GET"},
-		{"p", "888", "/sysDictionary/getSysDictionaryList", "GET"},
-		{"p", "888", "/sysOperationRecord/createSysOperationRecord", "POST"},
-		{"p", "888", "/sysOperationRecord/deleteSysOperationRecord", "DELETE"},
-		{"p", "888", "/sysOperationRecord/updateSysOperationRecord", "PUT"},
-		{"p", "888", "/sysOperationRecord/findSysOperationRecord", "GET"},
-		{"p", "888", "/sysOperationRecord/getSysOperationRecordList", "GET"},
-		{"p", "888", "/sysOperationRecord/deleteSysOperationRecordByIds", "DELETE"},
-		{"p", "888", "/user/setUserInfo", "PUT"},
-		{"p", "8881", "/base/login", "POST"},
-		{"p", "8881", "/base/register", "POST"},
-		{"p", "8881", "/api/createApi", "POST"},
-		{"p", "8881", "/api/getApiList", "POST"},
-		{"p", "8881", "/api/getApiById", "POST"},
-		{"p", "8881", "/api/deleteApi", "POST"},
-		{"p", "8881", "/api/updateApi", "POST"},
-		{"p", "8881", "/api/getAllApis", "POST"},
-		{"p", "8881", "/authority/createAuthority", "POST"},
-		{"p", "8881", "/authority/deleteAuthority", "POST"},
-		{"p", "8881", "/authority/getAuthorityList", "POST"},
-		{"p", "8881", "/authority/setDataAuthority", "POST"},
-		{"p", "8881", "/menu/getMenu", "POST"},
-		{"p", "8881", "/menu/getMenuList", "POST"},
-		{"p", "8881", "/menu/addBaseMenu", "POST"},
-		{"p", "8881", "/menu/getBaseMenuTree", "POST"},
-		{"p", "8881", "/menu/addMenuAuthority", "POST"},
-		{"p", "8881", "/menu/getMenuAuthority", "POST"},
-		{"p", "8881", "/menu/deleteBaseMenu", "POST"},
-		{"p", "8881", "/menu/updateBaseMenu", "POST"},
-		{"p", "8881", "/menu/getBaseMenuById", "POST"},
-		{"p", "8881", "/user/changePassword", "POST"},
-		{"p", "8881", "/user/getUserList", "POST"},
-		{"p", "8881", "/user/setUserAuthority", "POST"},
-		{"p", "8881", "/fileUploadAndDownload/upload", "POST"},
-		{"p", "8881", "/fileUploadAndDownload/getFileList", "POST"},
-		{"p", "8881", "/fileUploadAndDownload/deleteFile", "POST"},
-		{"p", "8881", "/casbin/updateCasbin", "POST"},
-		{"p", "8881", "/casbin/getPolicyPathByAuthorityId", "POST"},
-		{"p", "8881", "/jwt/jsonInBlacklist", "POST"},
-		{"p", "8881", "/system/getSystemConfig", "POST"},
-		{"p", "8881", "/system/setSystemConfig", "POST"},
-		{"p", "8881", "/customer/customer", "POST"},
-		{"p", "8881", "/customer/customer", "PUT"},
-		{"p", "8881", "/customer/customer", "DELETE"},
-		{"p", "8881", "/customer/customer", "GET"},
-		{"p", "8881", "/customer/customerList", "GET"},
-		{"p", "9528", "/base/login", "POST"},
-		{"p", "9528", "/base/register", "POST"},
-		{"p", "9528", "/api/createApi", "POST"},
-		{"p", "9528", "/api/getApiList", "POST"},
-		{"p", "9528", "/api/getApiById", "POST"},
-		{"p", "9528", "/api/deleteApi", "POST"},
-		{"p", "9528", "/api/updateApi", "POST"},
-		{"p", "9528", "/api/getAllApis", "POST"},
-		{"p", "9528", "/authority/createAuthority", "POST"},
-		{"p", "9528", "/authority/deleteAuthority", "POST"},
-		{"p", "9528", "/authority/getAuthorityList", "POST"},
-		{"p", "9528", "/authority/setDataAuthority", "POST"},
-		{"p", "9528", "/menu/getMenu", "POST"},
-		{"p", "9528", "/menu/getMenuList", "POST"},
-		{"p", "9528", "/menu/addBaseMenu", "POST"},
-		{"p", "9528", "/menu/getBaseMenuTree", "POST"},
-		{"p", "9528", "/menu/addMenuAuthority", "POST"},
-		{"p", "9528", "/menu/getMenuAuthority", "POST"},
-		{"p", "9528", "/menu/deleteBaseMenu", "POST"},
-		{"p", "9528", "/menu/updateBaseMenu", "POST"},
-		{"p", "9528", "/menu/getBaseMenuById", "POST"},
-		{"p", "9528", "/user/changePassword", "POST"},
-		{"p", "9528", "/user/getUserList", "POST"},
-		{"p", "9528", "/user/setUserAuthority", "POST"},
-		{"p", "9528", "/fileUploadAndDownload/upload", "POST"},
-		{"p", "9528", "/fileUploadAndDownload/getFileList", "POST"},
-		{"p", "9528", "/fileUploadAndDownload/deleteFile", "POST"},
-		{"p", "9528", "/casbin/updateCasbin", "POST"},
-		{"p", "9528", "/casbin/getPolicyPathByAuthorityId", "POST"},
-		{"p", "9528", "/jwt/jsonInBlacklist", "POST"},
-		{"p", "9528", "/system/getSystemConfig", "POST"},
-		{"p", "9528", "/system/setSystemConfig", "POST"},
-		{"p", "9528", "/customer/customer", "POST"},
-		{"p", "9528", "/customer/customer", "PUT"},
-		{"p", "9528", "/customer/customer", "DELETE"},
-		{"p", "9528", "/customer/customer", "GET"},
-		{"p", "9528", "/customer/customerList", "GET"},
-		{"p", "9528", "/autoCode/createTemp", "POST"},
-	}
-}
